% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_adultbmchange}
\alias{calc_adultbmchange}
\title{Update adult body mass}
\usage{
calc_adultbmchange(alive, BM_adult, Egain_adult, Ereq_adult, adult_mass_KG)
}
\arguments{
\item{alive}{Is the bird dead or alive? FALSE, TRUE}

\item{BM_adult}{Body mass of the adult bird, g}

\item{Egain_adult}{Energy actually acquired in the time step, kJ}

\item{Ereq_adult}{The full energy requirement for this time step, kJ}

\item{adult_mass_KG}{Energy density of the adult bird tissue, kJ per gram}
}
\value{
A revised adult body mass
}
\description{
Calculate body mass change. All adult birds update their body
mass at the end of each day based on the energy they gain and expend
foraging and in other activities. The model we use is an expanded version
of that used in Daunt & Wanless (2008) and Wanless et al. (1997), which
separates flight cost and foraging cost for each adult to derive total
energy expenditure
}
