% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_adultdee}
\alias{calc_adultdee}
\title{Calculate Daily Energy Requirement}
\usage{
calc_adultdee(
  alive,
  colony_h,
  flying_h,
  foraging_h,
  at_sea_h,
  energy_nest,
  energy_flight,
  energy_forage,
  energy_searest,
  energy_warming,
  assim_eff,
  daylength
)
}
\arguments{
\item{alive}{Is the bird dead or alive? 0,1}

\item{colony_h}{Time spent at the colony, hours}

\item{flying_h}{Time spent flying, hours}

\item{foraging_h}{Time spent foraging, hours}

\item{at_sea_h}{Time spent resting at sea, hours}

\item{energy_nest}{Energy cost of nesting at colony, kJ per day}

\item{energy_flight}{Energy cost of flight, kJ per day}

\item{energy_forage}{Energy cost of foraging,  kJ per day}

\item{energy_searest}{Energy cost of resting at sea, kJ per day}

\item{energy_warming}{Energy cost of warming food, kJ per day}

\item{assim_eff}{Assimilation efficiency}

\item{daylength, }{Length of this species' time step, hours}
}
\value{
A revised adult DEE for time step
}
\description{
This function calculates the energy expenditure timestep t for
adult birds, based on the activities carried out. This value is assumed to
be the energy requirement for the following time step, t+1.

DEE is the sum of proportion of total deployment time spent on each of these
activities multiplied by activity-specific energetic costs available from
the literature (Pennycuick 1987, 1989; Croll & McLaren 1993; Hilton et
al.2000a; Enstipp et al. 2006; respective costs: 1168.91 kJ day-1; 7361.72kJ
day-1; 810.28 kJ day-1; 1894.90 kJ day-1) and the cost of warming food
(Gremillet et al. 2003).
}
