% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions_distancedecay.R
\name{calc_birddensmap_dd_pinfr}
\alias{calc_birddensmap_dd_pinfr}
\title{Calculate bird density maps}
\usage{
calc_birddensmap_dd_pinfr(dmap, fr, pinfr, dmin = 1)
}
\arguments{
\item{dmap}{A raster containing the distance by sea from each grid cell to the population of interest}

\item{fr}{Foraging range, in kilometres}

\item{pinfr}{The proportion of the UD that is assumed to lie within a distance \code{fr} of the population: a numeric value between 0 and 1}
}
\value{
A raster containing the distance-decay map
\subsection{20 Oct 2023: added "dmin" argument}{
}
}
\description{
Calculate bird density maps using the distance-decay function used in SeabORD, via users specifying "proportion of UD lying within the foraging range"
}
