% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions_distancedecay.R
\name{calc_birddensmap_dd_pinhalf}
\alias{calc_birddensmap_dd_pinhalf}
\title{Calculate bird density maps used the distance-decay function used in SeabORD, via users specifying \code{phalf}}
\usage{
calc_birddensmap_dd_pinhalf(dmap, fr, pinhalf)
}
\arguments{
\item{dmap}{A raster containing the distance by sea from each grid cell to the population of interest}

\item{fr}{Foraging range, in kilometres}

\item{pinhalf}{the proportion of the foraging range within which half of the UD lies}
}
\value{
A raster containing the distance-decay map
}
\description{
Calculate bird density maps used the distance-decay function used in SeabORD, via users specifying \code{phalf}
}
