% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_chickcare}
\alias{calc_chickcare}
\title{Parenting the chicks}
\usage{
calc_chickcare(parent, ChickState, Opt_BM_chick, Species)
}
\arguments{
\item{parent}{A tibble containing the required variables from the adult birds
(BirdID, PairID, feeding_mode, Ereq_intakef_c, colony_h, MF).}

\item{ChickState}{List holding the current state of all the chicks and
metadata.}

\item{Opt_BM_chick}{The optimum body mass for the chick at this stage of the
season (a chick that has received full requirements every day).}

\item{Species}{List holding the species-specific variables (uses daylength,
BM_Chick_mortf and chick_mass_a)}
}
\value{
The updated list holding the tibbles for the chick current state.
}
\description{
For each chick, get the food provided by each of the parent
birds and increase the chick body mass. Compare with the 'optimum chick' to
estimate relative body condition and update the chick state. Add up
the total hours of attendance for the timestep.
}
