% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions_distancedecay.R
\name{calc_dd_phalf_from_pinfr}
\alias{calc_dd_phalf_from_pinfr}
\title{Calculate 'q' from 'p' for the distance-decay model}
\usage{
calc_dd_phalf_from_pinfr(pinfr)
}
\arguments{
\item{pinfr}{A vector of numeric values containing the values of 'pinfr'}
}
\value{
A vector of length \code{length(pinfr)} containing the values of \code{phalf} associated with \code{pinfr}
}
\description{
Calculate 'q' from 'p' for the distance-decay model
}
\details{
"pinfr": the probability of being within the foraging range; "phalf": the proportion of the foraging range within which half of the UD lies
Mathematical derivation:
Define q to be such that: \verb{[1 - exp(-beta * fr * phalf)] / [1 - exp(-beta * fr)] = 1/2}
Since: \verb{beta = −log(1 − pinfr) / fr}
We can immediately derive: \code{phalf = log(1 - (pinfr/2)) / log(1 - pinfr)}
}
