% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions_distancedecay.R
\name{calc_dd_pinfr_from_phalf}
\alias{calc_dd_pinfr_from_phalf}
\title{Calculate 'pinfr' from 'phalf' for the distance-decay model}
\usage{
calc_dd_pinfr_from_phalf(phalf)
}
\arguments{
\item{phalf}{A vector of numeric values containing the values of 'phalf'}
}
\value{
A vector of length \code{length(q)} containing the values of \code{pinfr} associated with \code{phalf}
}
\description{
Calculate 'pinfr' from 'phalf' for the distance-decay model
}
\details{
"pinfr": the probability of being within the foraging range; "phalf": the proportion of the UD within the foraging range that is within half the foraging range
This is optimized numerically
}
