% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions_distancedecay.R
\name{calc_dist_restricted}
\alias{calc_dist_restricted}
\title{Calculate distance restricted by obstructions}
\usage{
calc_dist_restricted(
  mymap,
  obspolys = NULL,
  targetcoords,
  FlightGridcorrection_3035,
  directions = 16,
  obspenalty = 1e+08,
  maxdist = 9000
)
}
\arguments{
\item{mymap}{A raster in which grid cells on land are assumed to have missing (NA) values}

\item{obspolys}{Polygons associated with obstacles; if NULL there are assumed to be no obstacles other than land}

\item{FlightGridcorrection_3035}{The 'flight correction' layer required by gdistance to correct for latitude (see gdistance)}

\item{obspenalty}{Penalty value associated with crossing an obstruction; a positive number}

\item{maxdist}{Maximum distance, in km; above this value are fixed to zero}
}
\value{
Raster, containing distance to target from each grid, avoiding obstructions, in kilometres
}
\description{
Calculate distance restricted by obstructions
}
\details{
Grid cells are treated as obstructed if (a) they have a missing (NA) value in the map, or, if "obspolys" has been provided, the grid cell lies within any of the polygons in "obspolys". Created 26 May 2024 based on "make_fltdist_scen" in SeabORD, but (a) for a single colony and (b) for obstruction either by land or other (e.g. footprints)
}
