% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_othermortality}
\alias{calc_othermortality}
\title{Chick mortality as a result of other causes}
\usage{
calc_othermortality(CoD, alive, seasonlength)
}
\arguments{
\item{CoD}{Cause of death  ("killed", "starved", "unattended", "parentdead", "abandoned", "other")}

\item{alive}{Flag to indicate if the chick is currently alive (1) or dead (0)}

\item{seasonlength}{Species parameter, number of timesteps in the breeding season}
}
\value{
A revised chick state list.
}
\description{
Chicks can be lost from other causes such as flooding, storms
etc. This function calculates the probability of death and then uses rbinom
to remove some chicks according to that probability.
}
