% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_puffinmortality}
\alias{calc_puffinmortality}
\title{Puffin chick mortality from predation due to hunger}
\usage{
calc_puffinmortality(CoD, alive, BM_condition, BM_Chick_mortf)
}
\arguments{
\item{CoD}{Cause of death ("killed", "starved", "unattended", "parentdead", "abandoned", "other")}

\item{alive}{Flag to indicate if the chick is currently alive (1) or dead (0)}

\item{BM_condition}{Body mass relative condition, 0-1}

\item{BM_Chick_mortf}{Critical mass below which chick is dead, 0-1}
}
\value{
A revised cause of death
}
\description{
If a puffin chick is hungry, there is an increased risk of
predation. This function takes the 'hungry' chicks (based on relative body
condition), calculates the probability of death by predation and then uses
rbinom to remove some of those chicks.
}
