% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_unattendmortality}
\alias{calc_unattendmortality}
\title{Chick mortality as a result of adults not attending the nest}
\usage{
calc_unattendmortality(CoD, alive, unattend_hrs, max_hrs)
}
\arguments{
\item{CoD}{Cause of death}

\item{alive}{Flag to indicate if the chick is currently alive (1) or dead (0)}

\item{unattend_hrs}{Number of hrs the chick was left unattended}

\item{max_hrs}{Species parameter: Critical time threshold for unattendance at
nest above which a chick is assumed to die through exposure or predation}
}
\value{
A revised cause of death
}
\description{
If a chick is left alone at the nest, there is an increased risk
of death from predation etc. This function takes the chicks who are not
fully attended by the adults, calculates the probability of death and then
uses rbinom to remove some of those chicks.
}
