% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cef_coast_3035}
\alias{cef_coast_3035}
\title{cef_coast_3035}
\format{
NEED EXPL.
\describe{
\item{COUNTRY }{Character: NEED EXPL.}
\item{ID_0}{Character: NEED EXPL}
\item{geometry}{Numeric: NEED EXPL}
}
}
\source{
NEED EXPL
}
\usage{
cef_coast_3035
}
\description{
NEED EXPL
}
\keyword{datasets}
