% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CEF_colsize_table}
\alias{CEF_colsize_table}
\title{CEF Size Table}
\format{
A data frame with 5 columns:
\describe{
\item{X}{Numeric: An identifier or index for the row.}
\item{species}{Character: The species code.}
\item{site_codes}{Character: The unique code identifying each site.}
\item{site_names}{Character: The name of the site.}
\item{spacolsize}{Numeric: NEED EXPL}
}
}
\source{
NEED EXPL
}
\usage{
CEF_colsize_table
}
\description{
Needs a description
}
\keyword{datasets}
