% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DBS_map_example}
\alias{DBS_map_example}
\title{DBS_map_example.R}
\format{
list of "raster" with the values of a raster in a matrix; and "metadata" with the extent and resolution to build back the raster in memory.
\describe{
\item{matrix}{matrix of pixel values}
\item{metadata}{List of metadata: extent, resolution, and CRS.}
}
}
\source{
Generated from "DBS_map.tif".
}
\usage{
DBS_map_example
}
\description{
NEED EXPL
}
\keyword{datasets}
