% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{energeticsandpreydata}
\alias{energeticsandpreydata}
\title{Energetics and Prey Data}
\format{
A data frame with the following 27 columns:
\describe{
\item{Code}{Character: A species code.}
\item{Species}{Character: The common name of the species.}
\item{BM_adult_mn}{Numeric: The mean body mass of adults.}
\item{BM_adult_sd}{Numeric: The standard deviation of adult body mass.}
\item{BM_adult_mortf}{Numeric: The mortality factor for adults.}
\item{BM_adult_abdn}{Numeric: The adult abandonment rate.}
\item{BM_chick_mn}{Numeric: The mean body mass of chicks.}
\item{BM_chick_sd}{Numeric: The standard deviation of chick body mass.}
\item{BM_Chick_mortf}{Numeric: The mortality factor for chicks.}
\item{daylength}{Numeric: The day length in hours.}
\item{seasonlength}{Numeric: The length of the season in days.}
\item{unattend_max_hrs}{Numeric: The maximum hours of unattended activity.}
\item{adult_DEE_mn}{Numeric: The mean daily energy expenditure (DEE) for adults.}
\item{adult_DEE_sd}{Numeric: The standard deviation of adult daily energy expenditure.}
\item{chick_DER}{Numeric: The chick daily energy requirement (DER).}
\item{IR_max}{Numeric: The maximum infrared radiation.}
\item{IR_half_a}{Numeric: Parameter related to infrared radiation.}
\item{IR_half_b}{Numeric: Another parameter related to infrared radiation.}
\item{flight_msec}{Numeric: The flight metabolic rate in milliseconds.}
\item{assim_eff}{Numeric: The assimilation efficiency.}
\item{energy_prey}{Numeric: The energy available from prey.}
\item{energy_nest}{Numeric: The energy associated with the nest.}
\item{energy_flight}{Numeric: The energy used in flight.}
\item{energy_searest}{Numeric: The energy associated with sea rest.}
\item{energy_forage}{Numeric: The energy used for foraging.}
\item{energy_warming}{Numeric: The energy used for warming.}
\item{chick_mass_a}{Numeric: The mass of chicks at age 'a'.}
\item{adult_mass_KG}{Numeric: The adult mass in kilograms.}
\item{beta}{Numeric: A parameter used in energetic calculations.}
\item{basesurv_poor}{Numeric: The survival rate under poor conditions.}
\item{basesurv_modr}{Numeric: The survival rate under moderate conditions.}
\item{basesurv_good}{Numeric: The survival rate under good conditions.}
\item{massloss_poor}{Numeric: The mass loss under poor conditions.}
\item{massloss_modr}{Numeric: The mass loss under moderate conditions.}
\item{massloss_good}{Numeric: The mass loss under good conditions.}
\item{chicksurv_poor}{Numeric: The chick survival rate under poor conditions.}
\item{chicksurv_modr}{Numeric: The chick survival rate under moderate conditions.}
\item{chicksurv_good}{Numeric: The chick survival rate under good conditions.}
}
}
\source{
NEED EXPL
}
\usage{
energeticsandpreydata
}
\description{
A dataset containing information on various bird species' energetics and prey data, including metabolic rates, energy requirements, and survival rates under different conditions.
}
\keyword{datasets}
