% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_1_lists}
\alias{example_1_lists}
\title{example_1_lists}
\format{
A list NEED EXPL
\describe{
\item{\code{switches}}{A list NEED EXPL}
\item{\code{Par}}{A list NEED EXPL}
\item{\code{modPar}}{A list NEED EXPL}
\item{\code{ordPar}}{A list NEED EXPL}
}
}
\source{
NEED EXPL
}
\usage{
example_1_lists
}
\description{
NEED EXPL
}
\keyword{datasets}
