% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{example_calibration_output}
\alias{example_calibration_output}
\title{example_calibration_output}
\format{
An object of class \code{data.frame} (or any class the \code{.rds} file contains).
\describe{
\item{switches}{list NEED EXPL}
\item{Parameters}{list NEED EXPL}
\item{ordPar}{list NEED EXPL}
\item{modPar}{list NEED EXPL}
\item{thisRun}{list NEED EXPL}
\item{output_f0}{list of tibbles}
\item{output_c0}{list of tibbles}
\item{output_a0}{list of tibbles}
\item{output_y0}{list of tibbles}
\item{birdflightamp}{raster}
}
}
\source{
\href{https://example.com}{Example Source}
}
\usage{
example_calibration_output
}
\description{
This dataset is stored in an \code{.rds} file NEED EXPL
}
\keyword{datasets}
