% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{frgcompdata_example}
\alias{frgcompdata_example}
\title{frgcompdata_example}
\format{
list of "raster" with the values of a raster in a matrix; and "metadata" with the extent and resolution to build back the raster in memory.
\describe{
\item{matrix}{matrix of pixel values}
\item{metadata}{List of metadata: extent, resolution, and CRS.}
}
}
\source{
NEED EXPL
}
\usage{
frgcompdata_example
}
\description{
NEED EXPL
}
\keyword{datasets}
