% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{memoised_calc_foragecapture}
\alias{memoised_calc_foragecapture}
\title{Memoized Version of calc_foragecapture}
\usage{
memoised_calc_foragecapture(...)
}
\description{
A memoized version of \code{\link{calc_foragecapture}} using \code{\link[memoise]{memoise}}.
This version caches results for improved performance on repeated calls with the same arguments.
}
\seealso{
\code{\link{calc_foragecapture}}, \code{\link[memoise]{memoise}}
}
