% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{offshorerenewabledevelopmentnames}
\alias{offshorerenewabledevelopmentnames}
\title{offshorerenewabledevelopmentnames}
\format{
A data frame with 5 columns:
\describe{
\item{code}{Character: NEED EXPL}
\item{name}{Character: NEED EXPL}
\item{Emodnet_Site_Name}{Character: NEED EXPL }
\item{Emodnet_PolygonID}{Numeric: NEED EXPL}
\item{Emodnet_Status}{Character: NEED EXPL}
}
}
\source{
NEED EXPL
}
\usage{
offshorerenewabledevelopmentnames
}
\description{
NEED EXPL
}
\keyword{datasets}
