% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ORDpoly_example_wfold}
\alias{ORDpoly_example_wfold}
\title{ORDpoly_example_wfold}
\format{
tibble
\describe{
\item{COUNTRY}{NEED EXPL}
\item{NAME}{NEED EXPL}
\item{N_TURBINES}{NEED EXPL}
\item{POWER_MW}{NEED EXPL}
\item{STATUS}{NEED EXPL}
\item{YEAR}{NEED EXPL}
\item{COAST_DIST_M}{NEED EXPL}
\item{AREA_SQKM}{NEED EXPL}
\item{NOTES}{NEED EXPL}
\item{Shape_Length}{NEED EXPL}
\item{Shape_Area}{NEED EXPL}
}
}
\source{
NEED EXPL
}
\usage{
ORDpoly_example_wfold
}
\description{
NEED EXPL
same projection as seamask
}
\keyword{datasets}
