% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordmain.R
\name{seabord}
\alias{seabord}
\title{SeabORD main function}
\usage{
seabord(
  Par,
  modPar,
  ordPar,
  switches,
  seamask,
  spadat1,
  spadat2,
  spdat,
  BrdData,
  FrgCompData,
  fltdist_base,
  FlightGridcorrection,
  ORDpoly
)
}
\arguments{
\item{Par}{List - The main parameters controlling/defining this run}

\item{modPar}{List - Parameters relating to the model mode and computer environment}

\item{ordPar}{List -  Input parameters relating to the ORDs}

\item{switches}{List -  A set of switches/flags used to control optional features of the run}

\item{seamask}{Raster - land/sea grid, 1km expected.}

\item{spadat1}{description TBC}

\item{spadat2}{description TBC}

\item{spdat}{Species-specific parameters}

\item{BrdData}{description TBC}

\item{FrgCompData}{description TBC}

\item{fltdist_base}{Flight distance by sea, without ORDs. See user guide for details.}

\item{FlightGridcorrection}{Flight distance transition layer (gdistance)}

\item{ORDpoly}{The ORD footprints}
}
\value{
A list containing tibbles
}
\description{
A model to estimate the population consequences of displacement
from proposed offshore renewable energy developments for key seabird species
}
