% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordsetup.R
\name{set_initialbirdstate}
\alias{set_initialbirdstate}
\title{Set initial values for individual birds}
\usage{
set_initialbirdstate(
  seed,
  N,
  tstep,
  BM_adult_mn,
  BM_adult_sd,
  adult_DEE_mn,
  adult_DEE_sd,
  assim_eff
)
}
\arguments{
\item{seed}{A seed to ensure reproducibility (numeric).}

\item{N}{The number of birds to generate (numeric).}

\item{tstep}{The length of the timestep in hours (numeric).}

\item{BM_adult_mn}{Adult body mass, mean, g (numeric).}

\item{BM_adult_sd}{Adult body mass, standard deviation (numeric).}

\item{adult_DEE_mn}{Adult daily energy expenditure, mean, kJ (numeric).}

\item{adult_DEE_sd}{Adult daily energy expenditure, standard deviation, kJ (numeric).}

\item{assim_eff}{Assimilation efficiency (numeric).}
}
\value{
A list containing a tibble holding the individual bird states at time
t0 and a tibble with metadata for the documentation.
}
\description{
Set the initial values for the birds in the simulation using
species parameters. All birds begin the simulation alive, part of a pair
with a live chick and in good condition.
}
\examples{
set_initialbirdstate(123, 10, 24, 372.7, 33.6, 802, 196, 0.74)

}
