% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordsetup.R
\name{set_initialbirdtype}
\alias{set_initialbirdtype}
\title{Create individual seabirds for the simulation}
\usage{
set_initialbirdtype(seed, Colonies, thisSpecies, Pwfde, Pwfbe)
}
\arguments{
\item{seed}{The seed used to ensure results are reproducible.}

\item{Colonies}{A tibble containing information about the colonies of
interest in this simulation. Requires "code" with the colony codes and
"ObsNpairs"}

\item{thisSpecies}{The two-letter code identifying the species of interest
(only works with one species in v2.0.0).}

\item{Pwfde}{The probability of displacement across the population if
encountering an ORD,where 0 means no displacement.}

\item{Pwfbe}{The probability across the population of experiencing a barrier
effect when encountering an ORD. Only applies to birds who are displacement
susceptible.}
}
\value{
A list containing two tibbles: the tidy data and the metadata for
documentation.
}
\description{
This function creates a tibble holding one row per individual
seabird. The values stored in this tibble do not change through the
simulation, they describe characteristics that remain constant like species,
sex and home colony.
}
\examples{
set_initialbirdtype(231, tibble::tibble(code="UK9004171", ObsNpairs=100), "Kw", 0.3, 1.0)

}
