% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordsetup.R
\name{set_initialchickstate}
\alias{set_initialchickstate}
\title{Set initial values for individual chicks}
\usage{
set_initialchickstate(
  seed,
  N,
  tstep,
  BM_chick_mn,
  BM_chick_sd,
  chick_DER,
  assim_eff
)
}
\arguments{
\item{seed}{A seed to ensure reproducibility (numeric).}

\item{N}{The number of chicks to generate (numeric).}

\item{tstep}{The length of the time step in hours (numeric).}

\item{BM_chick_mn}{Chick body mass, mean, g (numeric).}

\item{BM_chick_sd}{Chick body mass, standard deviation (numeric).}

\item{chick_DER}{Chick daily energy requirement, mean, kJ (numeric).}

\item{assim_eff}{Assimilation efficiency (numeric).}
}
\value{
A list containing two tibbles; one holding the individual chick
states at time t0 and the metadata for the documentation.
}
\description{
Set the initial values for the chicks in the simulation using
species parameters. All chicks begin the season alive and in good
condition.
}
\examples{
set_initialchickstate(123, 10, 24.0, 36.0, 2.2, 525.7, 0.74)

}
