% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordsetup.R
\name{set_medianprey}
\alias{set_medianprey}
\title{Set the median prey value across the region}
\usage{
set_medianprey(seed, PmaxLim, Nreplicates)
}
\arguments{
\item{seed}{The seed used to ensure results are reproducible (numeric).}

\item{PmaxLim}{The minimum and maximum allowed prey values (numeric).}

\item{Nreplicates}{The number of pairs of seasonal runs in the simulation (numeric).}
}
\value{
The median prey values to use in each pair of simulations.
}
\description{
SeabORD uses either uniformly distributed or mapped, variable
prey levels. If uniform, the same value for prey density is used for each
modelled cell in the region. If mapped, there is a median value used to
scale the prey availability. This function generates the median value for
each pair of seasonal runs in the simulation, between two user-defined
values.
}
\examples{
set_medianprey(3427, c(150, 350), 20)

}
