% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordsetup.R
\name{set_seedvalues}
\alias{set_seedvalues}
\title{Set the seeds needed for reproducibility}
\usage{
set_seedvalues(initialseed, nstreams, Nreplicates)
}
\arguments{
\item{initialseed}{The user-supplied main seed.}

\item{nstreams}{The number of different streams required.}

\item{Nreplicates}{The number simulation pairs in the run.}
}
\value{
A tibble with integer seed values for each stream and replicate
simulation pair.
}
\description{
SeabORD includes stochastic methods that rely on the generation
of random numbers. To make the simulations reproducible we start the random
number sequence using a seed. This function used the main, user-supplied seed
to create a table of other seeds that are used to initialise other functions.
Each stream in the simulation uses a separate seed so that, for example,
changes to the number of live birds does not affect the sequence in another
part of the code. Adding new streams or changing the number of replicates
should not change previous runs significantly (assuming the same starter
seed).
}
\examples{
  set_seedvalues(2327, 10, 5)
  
}
