% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spacoordinates}
\alias{spacoordinates}
\title{Spatial Coordinates of Sites}
\format{
A data frame with 23 columns:
\describe{
\item{SITECODE}{Character: The unique code identifying each site.}
\item{CEF.include}{Logical: Whether the site is included in the CEF dataset.}
\item{Marine}{Logical: Whether the site is marine.}
\item{Longitude}{Numeric: The longitude of the site.}
\item{Latitude}{Numeric: The latitude of the site.}
\item{dat.CELLNO}{Numeric: The cell number for the site in the data.}
\item{dat.ATSEA}{Logical: Whether the site is at sea.}
\item{flt.LONG}{Numeric: The longitude of the flight path.}
\item{flt.LAT}{Numeric: The latitude of the flight path.}
\item{flt.CELLNO}{Numeric: The cell number for the flight path.}
\item{flt.ATSEA}{Logical: Whether the flight path is at sea.}
\item{flt.DIST}{Numeric: The distance of the flight path from the site.}
\item{flt.near}{Logical: Whether the flight path is near the site.}
\item{datxy.E}{Numeric: The x-coordinate in the East direction for the site.}
\item{datxy.N}{Numeric: The y-coordinate in the North direction for the site.}
\item{datxy.CELLNO}{Numeric: The cell number for the site in the XY grid.}
\item{datxy.ATSEA}{Logical: Whether the XY coordinates are at sea.}
\item{fltxy.E}{Numeric: The x-coordinate in the East direction for the flight path.}
\item{fltxy.N}{Numeric: The y-coordinate in the North direction for the flight path.}
\item{fltxy.CELLNO}{Numeric: The cell number for the flight path in the XY grid.}
\item{fltxy.ATSEA}{Logical: Whether the XY coordinates for the flight path are at sea.}
\item{fltxy.DIST}{Numeric: The distance of the flight path in XY space.}
\item{fltxy.near}{Logical: Whether the flight path is near the site in XY space.}
}
}
\source{
NEED EXPL
}
\usage{
spacoordinates
}
\description{
A dataset containing spatial coordinates and related attributes for various sites, including information on whether the site is marine, whether data is available for different columns, and the spatial proximity of flight paths.
}
\keyword{datasets}
