% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{spalist}
\alias{spalist}
\title{SPA Site List Dataset}
\format{
A data frame with the following 10 columns:
\describe{
\item{SITE_CODE}{Character: A unique code identifying each SPA.}
\item{SITE_NAME}{Character: The name of the SPA.}
\item{CEF.include}{character: Indicates if the site is included in the CEF project (Y) or not (NA).}
\item{Marine.only}{character: Indicates if the site is classified as a marine-only SPA.}
\item{Altnames.sufficient}{character:NEED EXPL}
\item{Altnames.SPApolys}{character:NEED EXPL}
\item{Altnames.SMP}{character:NEED EXPL}
\item{Altnames.Productivity}{character:NEED EXPL}
\item{Altnames.BDMPS}{character:NEED EXPL}
\item{Altnames.FR}{character:NEED EXPL.}
}
}
\source{
NEED EXPL
}
\usage{
spalist
}
\description{
This dataset contains information about Special Protection Areas (SPAs), including site codes, site names.
}
\keyword{datasets}
