% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{specieslist}
\alias{specieslist}
\title{Species List Dataset}
\format{
A data frame with the following 13 columns:
\describe{
\item{ENGLISH_NAME}{Character: The English name of the species.}
\item{LATIN_NAME}{Character: The scientific (Latin) name of the species.}
\item{TAXA_CODE}{Character: A short taxa code associated with the species.}
\item{SHORT_CODE}{Character: A short code used for quick reference.}
\item{FIVE_LETTER}{Character: A five-letter shorthand for the species.}
\item{BOU_NAME}{Character: The name of the species as recognized by the British Ornithologists' Union (BOU).}
\item{sCRM_NAME}{Character:  NEED EXPL}
\item{CEF}{character: Indicates if the species is included in the CEF project.}
\item{iPCoD}{character: Indicates if the species is included in the iPCoD project.}
\item{SeabORD}{character: Indicates if the species is included in the SeabORD project.}
\item{sCRM}{character: Indicates if the species is included in the sCRM project.}
\item{NEPVA}{character: Indicates if the species is included in the NEPVA project.}
\item{DispMat}{character: Indicates if dispersal matrices are available for the species.}
}
}
\source{
NEED EXPL
}
\usage{
specieslist
}
\description{
This dataset provides information about the bird species included in seabORD, including their English and Latin names, taxa codes, and 2 letter codes.
}
\keyword{datasets}
