% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UK9004171_bysea_3035}
\alias{UK9004171_bysea_3035}
\title{UK9004171_bysea_3035}
\format{
list of "raster" with the values of a raster in a matrix; and "metadata" with the extent and resolution to build back the raster in memory.
\describe{
\item{matrix}{matrix of pixel values}
\item{metadata}{List of metadata: extent, resolution, and CRS.}
}
}
\source{
Generated from "UK9004171_bysea_3035.tif".
}
\usage{
UK9004171_bysea_3035
}
\description{
NEED EXPL
}
\keyword{datasets}
