% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions_distancedecay.R
\name{voptimise}
\alias{voptimise}
\title{Use numerical optimisation to give the value of \code{x} associated with each element of a vector \code{y}, where \code{y = fny(x)} for a specified function \code{fny}}
\usage{
voptimise(y, fny, iv)
}
\arguments{
\item{y}{A vector of numerical values}

\item{fny}{A function with a single argument \code{x} that produces the value of \code{y} associated with \code{x}}

\item{iv}{A vector of length two indicating the numerical range of values to optimize over}
}
\value{
A vector of numeric values of lengt \code{y} containing the value of \code{x} associated with each value of \code{y}
}
\description{
Use numerical optimisation to give the value of \code{x} associated with each element of a vector \code{y}, where \code{y = fny(x)} for a specified function \code{fny}
}
