% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-seabordbirds.R
\name{calc_pSurvival}
\alias{calc_pSurvival}
\title{Probability of winter survival (mass-survival)}
\usage{
calc_pSurvival(sp, bmi, bm, basesurv, beta, sd = 0)
}
\arguments{
\item{sp}{The species two-letter code (either from "KI", "GU", "KI" or "RA")}

\item{bmi}{Body mass for the individual adult bird (g)}

\item{bm}{Body mass of the adult birds alive at the end of the
breeding season, mean (g)}

\item{basesurv}{Baseline survival}

\item{beta}{Mass-survival slope}

\item{sd}{Body mass of the adult birds alive at the end of the
breeding season, standard deviation}
}
\value{
A probability of survival
}
\description{
Calculate the probability of survival over the whole year based
on the body mass of the individual relative to the population mean, species
expected survival and parameters. Baseline survival values may apply to
poor, moderate or good years.
}
\examples{
  calc_pSurvival("KI", 345.9, 370.8, 0.8, 0.038, 0)
  calc_pSurvival("GU", 345.9, 370.8, 0.92, 1.03, 50)
}
